requirements.txt
attrs==20.2.0
click==7.1.2
Flask==1.1.2
Flask-Injector==0.12.3
iniconfig==1.0.1
injector==0.18.3
itsdangerous==1.1.0
Jinja2==2.11.2
MarkupSafe==1.1.1
more-itertools==8.5.0
packaging==20.4
pluggy==0.13.1
py==1.9.0
pyparsing==2.4.7
pytest==6.0.2
python-dotenv==0.14.0
six==1.15.0
toml==0.10.1
typing-extensions==3.7.4.3
Werkzeug==1.0.1
